python_service/requirements.txt
openai>=1.12.0
faster-whisper>=1.1.1
ffmpeg-python
fastapi>=0.68.0
uvicorn>=0.15.0
torch>=2.0.0
numpy>=1.21.0
python-dotenv>=0.19.0
pytube
yt-dlp>=2023.7.6
psutil>=5.8.0
redis>=4.5.5
rq>=1.15.1
